# docs/swagger.yaml
definitions:
  model.UploadResponse:
    properties:
      content-length:
        type: integer
      etag:
        type: string
      last-modified:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/copy:
    post:
      consumes:
      - application/json
      description: 处理文件拷贝请求，支持在不同存储桶之间拷贝文件
      parameters:
      - default: os.Getenv("COS_BUCKET"
        description: 源存储桶
        in: query
        name: srcBucket
        type: string
      - description: 源文件对象名
        in: query
        name: srcObject
        required: true
        type: string
      - default: srcBucket
        description: 目标存储桶
        in: query
        name: destBucket
        type: string
      - description: 目标文件对象名
        in: query
        name: destObject
        required: true
        type: string
      - default: os.Getenv("COS_REGION"
        description: 源存储桶区域
        in: query
        name: srcRegion
        type: string
      - default: os.Getenv("COS_REGION"
        description: 目标存储桶区域
        in: query
        name: destRegion
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 文件拷贝成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 缺少必需的参数
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 文件拷贝失败
          schema:
            additionalProperties: true
            type: object
      summary: 拷贝文件
      tags:
      - 文件管理
  /api/v1/delete:
    delete:
      consumes:
      - application/json
      description: 处理文件删除请求
      parameters:
      - description: 文件对象名
        in: query
        name: objectName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 文件删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数缺失
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 文件删除失败
          schema:
            additionalProperties: true
            type: object
      summary: 删除文件
      tags:
      - 文件管理
  /api/v1/download:
    get:
      consumes:
      - application/json
      description: 处理文件下载请求
      parameters:
      - description: 文件对象名
        in: query
        name: object_Name
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 文件数据
          schema:
            type: file
        "400":
          description: 参数缺失或文件下载失败
          schema:
            additionalProperties: true
            type: object
      summary: 下载文件
      tags:
      - 文件管理
  /api/v1/list:
    get:
      consumes:
      - application/json
      description: 获取指定目录下的文件列表，并支持分页查询
      parameters:
      - description: 文件前缀
        in: query
        name: prefix
        type: string
      - description: 分页查询标记，继续上次查询的位置
        in: query
        name: marker
        type: string
      - description: 每页返回的文件数，最大值为1000，默认为1000
        in: query
        name: maxKeys
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 文件列表获取成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid maxKeys parameter
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 获取文件列表失败
          schema:
            additionalProperties: true
            type: object
      summary: 获取文件列表
      tags:
      - 文件管理
  /api/v1/move:
    post:
      consumes:
      - application/json
      description: 处理文件移动请求，支持在同一存储桶或不同存储桶之间移动文件
      parameters:
      - default: os.Getenv("COS_BUCKET"
        description: 源存储桶
        in: query
        name: srcBucket
        type: string
      - description: 源文件对象名
        in: query
        name: srcObject
        required: true
        type: string
      - default: srcBucket
        description: 目标存储桶
        in: query
        name: destBucket
        type: string
      - description: 目标文件对象名
        in: query
        name: destObject
        required: true
        type: string
      - default: os.Getenv("COS_REGION"
        description: 源存储桶区域
        in: query
        name: srcRegion
        type: string
      - default: os.Getenv("COS_REGION"
        description: 目标存储桶区域
        in: query
        name: destRegion
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 文件移动成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 缺少必需的参数
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 文件移动失败
          schema:
            additionalProperties: true
            type: object
      summary: 移动文件
      tags:
      - 文件管理
  /api/v1/upload:
    post:
      consumes:
      - multipart/form-data
      description: 处理文件上传请求
      parameters:
      - description: 上传的文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            $ref: '#/definitions/model.UploadResponse'
        "400":
          description: 上传失败
          schema:
            additionalProperties: true
            type: object
      summary: 上传文件
      tags:
      - 文件管理
swagger: "2.0"